+++ src/config.ts
@@ -120,12 +120,12 @@ export const DEFAULT_CONFIG: Config = {
     // Culled from https://mastodonservers.net and https://joinmastodon.org/
     defaultServers: [
         "mastodon.social",
-        // "pawoo.net",   // Japanese (and maybe NSFW?)
-        // "baraag.net",  // very NSFW
-        // "mstdn.jp",    // Japanese
+        // "pawoo.net",       // Japanese (and maybe NSFW?)
+        // "baraag.net",      // very NSFW
+        // "mstdn.jp",        // Japanese
         "mastodon.cloud",
-        // "pravda.me"    // Russian
-        "mstdn.social",
+        // "pravda.me"        // Russian
+        // "mstdn.social",    // Slow, blocked by CORS
         "mastodon.online",
         "mas.to",
         "mastodon.world",
@@ -158,6 +158,7 @@ export const DEFAULT_CONFIG: Config = {
         "fediverse.one",
         "flipboard.com",
         'mathstodon.xyz',
+        "mstdn.social",    // blocked by CORS
         "threads.net",
     ],
 };