+++ src/helpers.ts
@@ -6,8 +6,9 @@ import { mastodon } from "masto";
 
 import { MediaCategory, CountKey, StringNumberDict, Weights } from "./types";
 
+export const GIFV = "gifv";
 export const IMAGE_EXTENSIONS = ["jpg", "jpeg", "png"];
-export const VIDEO_TYPES = [MediaCategory.VIDEO, "gifv"] as mastodon.v1.MediaAttachmentType[];
+export const VIDEO_TYPES = [MediaCategory.VIDEO, GIFV] as mastodon.v1.MediaAttachmentType[];
 export const VIDEO_EXTENSIONS = ["mp4"];
 export const MEDIA_TYPES = [MediaCategory.AUDIO, MediaCategory.IMAGE, ...VIDEO_TYPES];
 export const DEFAULT_FONT_SIZE = 15;