+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # NEXT RELEASE
 * Refactor `MediaCategory` enum
+* Explicitly export all objects and variables required by the demo app
+* Exclude `mstd.social` from default servers
 
 ## 0.17.0
 * Configure various algorithm weight presets that can be selected via `PresetWeights` object